+++ admin/src/main/resources/io/buoyant/admin/js/main-namerd.js
@@ -1,0 +1,29 @@
+require.config({
+  paths: {
+    'jQuery': 'lib/jquery.min',
+    'lodash': 'lib/lodash.min',
+    'Handlebars': 'lib/handlebars-v4.0.5',
+    'bootstrap': 'lib/bootstrap.min',
+    'text': 'lib/text'
+  },
+  shim: {
+    'jQuery': {
+      exports: '$'
+    },
+    'lodash': {
+      exports: '_'
+    },
+    'bootstrap': {
+      deps : ['jQuery'],
+      exports: 'Bootstrap'
+    }
+  }
+});
+
+require([
+  'jQuery',
+  'bootstrap',
+  'src/dashboard_delegate'
+], function ($, bootstrap, namerdDtabPlayground) {
+  new namerdDtabPlayground();
+});