+++ linkerd/core/src/main/scala/io/buoyant/linkerd/AnnouncerInitializer.scala
@@ -3,6 +3,7 @@ package io.buoyant.linkerd
 import com.fasterxml.jackson.annotation.{JsonIgnore, JsonProperty, JsonTypeInfo}
 import com.twitter.finagle.Path
 import io.buoyant.config.ConfigInitializer
+import io.buoyant.namer.Paths
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "kind")
 trait AnnouncerConfig {
@@ -14,14 +15,10 @@ trait AnnouncerConfig {
   def defaultPrefix: Path
 
   @JsonIgnore
-  def prefix: Path = AnnouncerConfig.hash ++ _prefix.getOrElse(defaultPrefix)
+  def prefix: Path = Paths.ConfiguredNamerPrefix ++ _prefix.getOrElse(defaultPrefix)
 
   @JsonIgnore
   def mk(): Announcer
-}
-
-object AnnouncerConfig {
-  val hash = Path.Utf8("#")
 }
 
 trait AnnouncerInitializer extends ConfigInitializer