+++ linkerd/failure-accrual/src/main/scala/io/buoyant/linkerd/failureAccrual/NoneInitializer.scala
@@ -1,0 +1,24 @@
+package io.buoyant.linkerd.failureAccrual
+
+import com.fasterxml.jackson.annotation.JsonIgnore
+import com.twitter.finagle.service.exp.FailureAccrualPolicy
+import com.twitter.util.Duration
+import io.buoyant.linkerd.{FailureAccrualConfig, FailureAccrualInitializer}
+
+class NoneInitializer extends FailureAccrualInitializer {
+  val configClass = classOf[NoneConfig]
+  override def configId = "none"
+}
+
+object NoneInitializer extends NoneInitializer
+
+class NoneConfig extends FailureAccrualConfig {
+  @JsonIgnore
+  override def policy = () => NonePolicy
+}
+
+object NonePolicy extends FailureAccrualPolicy {
+  override def recordSuccess(): Unit = {}
+  override def markDeadOnFailure(): Option[Duration] = None
+  override def revived(): Unit = {}
+}